+++ tests/test_widgets.py
@@ -1,0 +1,39 @@
+import unittest
+
+from PySide6 import QtCore, QtWidgets
+
+from qtbricks import widgets
+
+
+class TestQHLine(unittest.TestCase):
+    def setUp(self):
+        self.app = (
+            QtWidgets.QApplication.instance() or QtWidgets.QApplication()
+        )
+        self.widget = widgets.QHLine()
+        self.addCleanup(self.release_qt_resources)
+
+    def release_qt_resources(self):
+        self.widget.deleteLater()
+        self.app.sendPostedEvents(event_type=QtCore.QEvent.DeferredDelete)
+        self.app.processEvents()
+
+    def test_instantiate_class(self):
+        pass
+
+
+class TestQVLine(unittest.TestCase):
+    def setUp(self):
+        self.app = (
+            QtWidgets.QApplication.instance() or QtWidgets.QApplication()
+        )
+        self.widget = widgets.QVLine()
+        self.addCleanup(self.release_qt_resources)
+
+    def release_qt_resources(self):
+        self.widget.deleteLater()
+        self.app.sendPostedEvents(event_type=QtCore.QEvent.DeferredDelete)
+        self.app.processEvents()
+
+    def test_instantiate_class(self):
+        pass