+++ docs/changelog.rst
@@ -5,8 +5,32 @@ Changelog
 This page contains a summary of changes between the official qtbricks releases. Only the biggest changes are listed here. A complete and detailed log of all changes is available through the `GitHub Repository Browser <https://github.com/tillbiskup/qtbricks>`_.
 
 
-Version 0.1.0-rc.2
-==================
+Version 0.1.0
+=============
+
+Not yet released
+
+* First public release
+
+* ...
+
+
+Version 0.1.0-rc3
+=================
+
+Released 2025-08-02
+
+* Third public pre-release
+
+
+New features
+------------
+
+* New module :mod:`qtbricks.widgets` containing small Qt widgets for general use with Qt GUIs.
+
+
+Version 0.1.0-rc2
+=================
 
 Released 2024-01-15